app/crawler/requirements.txt
attrs==23.1.0
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.2
cssselect==1.2.0
filelock==3.12.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
packaging==23.1
parsel==1.8.1
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
scrapy-splash==0.9.0
service-identity==23.1.0
six==1.16.0
tldextract==3.4.4
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==2.0.3
w3lib==2.1.1
zope.interface==6.0
attrs==23.1.0
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.2
cssselect==1.2.0
filelock==3.12.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
packaging==23.1
parsel==1.8.1
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
scrapy-splash==0.9.0
service-identity==23.1.0
six==1.16.0
tldextract==3.4.4
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==2.0.3
w3lib==2.1.1
zope.interface==6.0
alembic==1.11.1
annotated-types==0.5.0
dnspython==2.3.0
greenlet==2.0.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
pydantic==2.0.2
pydantic_core==2.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.18
SQLAlchemy-Utils==0.41.1
tenacity==8.2.2
typing_extensions==4.7.1
